fix: background is offset relative to markers
NEVER FORGET THE FIRST RULE OF WORKING WITH CHROMIUM-BASED BROWSERS:
SUBPIXELS ARE ROUNDED TO PIXELS. ALWAYS MULTIPLY SIZES BY A LARGE ENOUGH
NUMBER SO THIS DOES NOT BECOME A PROBLEM. THANKS GOOGLE!
Also, how is the same image can be rendered in different locations based
on its contents? I changed the images and it became offset.
It is also offset on Android.