make a plugin to load worker synchronously in <head>
Before that, there was a serial dependency chain on
index.html > render.js > worker.js > all the files.
Now the worker creation is inline in index.html and we get
index.html > worker.js > all the files. Would be ideal to
prefetch all the files as well, since there aren't much,
but prefetching for use in WebWorker is not possible in html