fix: copy url animation replays after switching menus
Remove the animation. Imagine standardizing that animations
replay every time the element becomes visible? And making it
impossible to play animations from JS the normal way. Why
make 'stateless' animations, when you can have animations
that are actually useful? Also now there's a dummy element
that displays success and fail color, because it is PHYSICALLY
IMPOSSIBLE to draw 2 colors on top of each other; we need two
different elements for that (also because :hover changes background).